backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
sqlalchemy[asyncio]>=2.0,<3.0
asyncpg>=0.30,<1.0
alembic>=1.14,<2.0
pydantic>=2.10,<3.0
pydantic-settings>=2.0,<3.0
celery>=5.4,<6.0
redis>=5.0,<6.0
gspread>=6.0,<7.0
google-auth>=2.0,<3.0
httpx>=0.28,<1.0
loguru>=0.7,<1.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0
pytest-asyncio>=0.24,<1.0
aiosqlite>=0.20,<1.0
